refactor system handle to break it down to few subclasses
